+++ src/Support/Grid.php
@@ -2,8 +2,4 @@
 
 namespace HkDevs\CodeForgeStudio\Support;
 
-if (class_exists(\Filament\Schemas\Components\Grid::class)) {
-    class Grid extends \Filament\Schemas\Components\Grid {}
-} else {
-    class Grid extends \Filament\Forms\Components\Grid {}
-}
+class Grid extends \Filament\Schemas\Components\Grid {}